requirements.txt
e3db~=2.1
psycopg2-binary~=2.7
boto3~=1.9